Ensure proper rendering of malformed SVGs
I noticed some of the country flags were rendering only a small portion of the SVG. Upon experimentation, I found that this was due to the missing viewBox property. After some digging I discovered a configuration option on SVGO that ensured viewBox was present using the SVGs width and height as a bases to construct the viewBox.
However, using the SVGO options overwrites the default options used by SVGR, so I had to re-add some of those crucial plugins back too.